requirements.txt
amadeus==8.1.0
annotated-types==0.7.0
anyio==4.10.0
attrs==23.2.0
Babel==2.10.3
bcc==0.1.10
blinker==1.7.0
Brlapi==0.8.5
cachetools==5.5.2
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
deprecation==2.1.0
distro==1.9.0
distro-info==1.7+build1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.116.2
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gotrue==2.12.4
griffe==1.14.0
gyp==0.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.27.2
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.6
Jinja2==3.1.2
jiter==0.11.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mcp==1.14.1
mdurl==0.1.2
msal==1.33.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
openai==1.108.1
openai-agents==0.3.1
packaging==25.0
pexpect==4.9.0
pillow==10.2.0
postgrest==0.17.2
proto-plus==1.26.1
protobuf==6.32.1
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.25.1
pycups==2.0.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.10.1
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
realtime==2.19.0
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.27.1
rsa==4.9.1
setuptools==68.1.2
six==1.16.0
sniffio==1.3.1
sse-starlette==3.0.2
ssh-import-id==5.11
starlette==0.48.0
storage3==0.8.2
stripe==12.5.1
supabase==2.9.1
supafunc==0.6.2
systemd-python==235
tenacity==9.0.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.0.7
uvicorn==0.36.0
wadllib==1.3.6
websockets==15.0.1
wheel==0.42.0
xdg==5
xkit==0.0.0
